Move observation location transformation into plot module
Represent plot observation locations with explicit coordinates
