+++ src/ert/field_utils/__init__.py
@@ -11,7 +11,6 @@ from .field_utils import (
     read_field,
     save_field,
     transform_local_ellipse_angle_to_local_coords,
-    transform_observation_locations,
     transform_positions_to_local_field_coordinates,
 )
 
@@ -26,6 +25,5 @@ __all__ = [
     "read_field",
     "save_field",
     "transform_local_ellipse_angle_to_local_coords",
-    "transform_observation_locations",
     "transform_positions_to_local_field_coordinates",
 ]
