+++ src/ert/gui/tools/plot/plottery/plots/cesp.py
@@ -18,6 +18,7 @@ if TYPE_CHECKING:
     from matplotlib.figure import Figure
     from pandas import DataFrame
 
+    from ert.gui.tools.plot.plot_types import ObservationPlotLocations
     from ert.gui.tools.plot.plottery import PlotConfig, PlotContext
 
 
@@ -46,7 +47,7 @@ class CrossEnsembleStatisticsPlot:
         ensemble_to_data_map: dict[EnsembleObject, pd.DataFrame],
         observation_data: pd.DataFrame,
         std_dev_images: dict[str, npt.NDArray[np.float32]],
-        obs_loc: npt.NDArray[np.float32] | None,
+        obs_loc: ObservationPlotLocations | None,
         key_def: PlotApiKeyDefinition | None = None,
     ) -> None:
         plotCrossEnsembleStatistics(
