+++ tests/ert/unit_tests/field_utils/test_field_utils_for_distance_based_localization.py
@@ -1,14 +1,10 @@
 import numpy as np
-import pandas as pd
 import pytest
 from numpy.testing import assert_allclose
 
 from ert.field_utils import (
-    AxisOrientation,
-    ErtboxParameters,
     calc_rho_for_2d_grid_layer,
     transform_local_ellipse_angle_to_local_coords,
-    transform_observation_locations,
     transform_positions_to_local_field_coordinates,
 )
 
@@ -91,70 +87,6 @@ def test_transform_positions_to_local_field_coordinates(
     )
     assert_allclose(transf_x, expected_x)
     assert_allclose(transf_y, expected_y)
-
-
-@pytest.mark.parametrize(
-    (
-        "east",
-        "north",
-        "ertbox_params",
-        "result_loc",
-    ),
-    [
-        pytest.param(
-            [1, 2],
-            [0, 0],
-            ErtboxParameters(
-                origin=(1000.0, 1000.0),
-                rotation_angle=0.0,
-                nx=10,
-                ny=10,
-                nz=10,
-                axis_orientation=AxisOrientation.LEFT_HANDED,
-            ),
-            None,
-            id="Observations outside the field (left-handed)",
-        ),
-        pytest.param(
-            [1, 2],
-            [2, 2],
-            ErtboxParameters(
-                origin=(1.0, 1.0),
-                rotation_angle=0.0,
-                nx=10,
-                ny=10,
-                nz=10,
-                axis_orientation=AxisOrientation.LEFT_HANDED,
-            ),
-            np.array([[0.0, 1.0], [1.0, 1.0]], dtype=np.float32),
-            id="Observations inside the field (left-handed)",
-        ),
-        pytest.param(
-            [1, 2],
-            [2, 2],
-            ErtboxParameters(
-                origin=(1.0, 1.0),
-                rotation_angle=0.0,
-                nx=10,
-                ny=10,
-                nz=10,
-                axis_orientation=AxisOrientation.RIGHT_HANDED,
-            ),
-            np.array([[0.0, 9.0], [1.0, 9.0]], dtype=np.float32),
-            id="Observations inside the field (right-handed)",
-        ),
-    ],
-)
-def test_that_transform_observation_locations_handles_different_cases(
-    east, north, ertbox_params, result_loc
-):
-
-    df = pd.DataFrame({"east": east, "north": north})
-    result = transform_observation_locations(df, ertbox_params)
-    if result is None:
-        assert result_loc is None
-    else:
-        assert np.array_equal(result, result_loc)
 
 
 def test_that_calc_rho_for_2d_grid_layer_ignores_obs_outside_the_grid():
